ICS browser tab settings changes
Allow the user to toggle whether or not browser tabs are restored on browser launch, either after killing the browser or after reboot. Allow the user to determine the maximum number of allowable tabs instead of using a hard-coded integer.

Patch-set 3: Changed the summary for the restore tabs setting to match the description in the commit.

Signed-off-by: Gary Danko <gdanko@gmail.com>
Change-Id: If594fdaf55f471f59fd597171f1bfe68ec384f6c