Browser tab settings.
Allow the user to toggle whether or not browser tabs are restored on
browser launch, either after killing the browser or after reboot.
Allow the user to determine the maximum number of allowable tabs
instead of using a hard-coded integer.

Patch-set 2: Remove the restore previous tabs for now. It doesn't
work in this version of Browsert. I will re-visit this later.

Patch-set 3: Tabs > spaces :/

Signed-off-by: gdanko <gdanko@gmail.com>